+++ src/main.js
@@ -1,5 +1,9 @@
 import Vue from 'vue';
 import App from './App.vue';
+import './assets/style.css';
+import EasterEggTrigger from './index';
+
+Vue.use(EasterEggTrigger);
 
 Vue.config.productionTip = false;