+++ src/plugin/index.js
@@ -1,0 +1,3 @@
+import EasterEggTriggerCore from './EasterEggTriggerCore';
+
+export default EasterEggTriggerCore;