+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log
 All notable changes to the "vue-easter-egg-trigger" plugin will be documented in this file.
 
-## v0.1.24
+## v1.0.0
 April 27, 2021
-[main] Initial release (webdevnerdstuff)
+ 
+Initial release
+
+## v2.0.0
+February 11, 2022
+ 
+1. Updating plugin to support both usage as a component as well as an instance method.
+2. Removed the `key` option as it was using depreciated key events.
+3. Fixed some minor bugs.
+4. Added a demo page.